Allows Vox defibrillation
Makes use of the IS_VOX reagent tag to allow Vox to be defibrillated without removing their NO_SCAN tag. Due to Vox having such a high shock resistance (80% base resistance, according to the Seimens Coefficient of 0.2), this feature is restricted to Defibrillators that can bypass armor, voidsuits, and RIG, such as the Paws of Life and Combat Defibrillator. Anything else will return a message declaring that the voltage rating is insufficient and abort the defib attempt.