Reduce Nif wear of Mass Alteration
a simple line change to make it 3 wear on the nif. It was already absurdly high compared to literally everything else in the nif shop. I brought it down and it will be discussed at the round table tomorrow. Since it's not meant for combat in the first place, the intense nif wear just felt strange.