V2/V4 3x high activity retry
Reenables a few old ruins that were removed 4no reason, aswell as removes a few that are either impractical, spawn over and over reguardless of not containting anything, or just dont work in the first place. Map inits are only increased by 50-80 seconds, compared to the previous 300 last time. I plan on adding new ruins later, but untill then, this 'should' be finished and ready to merge.