Adds an option to fix conveyor movement direction
For one reason or another, conveyors, when constructed by players, have south and north set as their forwards and backwards directions, despite the conveyor's actual direction. I couldn't find a way to just make it fix on spawning, but, I did the next best thing.
If you use a wrench on a conveyor that isn't operating and has its panel open, it runs the direction-setting process again, fixing the forwards and backwards directions.