fixes books
book code was using the config.wikiurl to get the wiki url for books, and as that is not and has never been set, i changed it to the proper full vorestation urls.

if i was able to set the wikiurl, i would have just set the wikiurl.

also i had to fix the ater_stairs changelog bc you forgot to put a colon after line 3 whoever you are.