Humans now have genders
previously all human bodies were male regardless of their gender, now the females have female parts and males have male parts. as of this PR, the preview screen in character setup will still default to male. however, in-game, this is not the case as they will be the appropriate gender based on their selection