Protean health tweaks
Tweaked: 	O.damage = max(0,O.damage-0.1) to 	O.damage = max(0,O.damage-1). Allows for proteans to heal 100 toxins without waiting forever and running out of metal.

Tweaked:	if(health < maxHealth*0.5) to 	if(health < maxHealth*0.35) to allow for protean to still be somewhat useful in combat (current forces proteans to auto-blob at half health)
The max health change should be reverted if the heal rate is not as slow as currently.