Fixes robots
Vanilla power charge is 7500, meaning that in around 5-10 minutes of joining they will have to seek a recharger. This is not becoming of an RP server, where a single emote can take this long to be sent! Switches their cell to the super cell, nearly 3x the capacity for 3x the emotes.