Unfucks the kitchen grill
The grill in the kitchen previously failed to check if it's empty when
emptied of all items. This is a small refactor of the grill's special
snowflake item removal proc and adds the check to see if it's empty
after taking out an item, and cleanly stops cooking if so.