Adds extra hair sprites
The duplicates in human_face_m and human_face need to be there to get the hairs to work, I kept the beards as well because I don't know if removing it there will fuck it up. The beards have to be in human_face_or_vr else they won't load right.