Adds Songs to the Jukebox
This update brings an abundance of music to the jukebox, so that
we can finally listen to something other than the same four songs on
repeat like we have been for who knows how long now. Additionally, this
update lays some better foundation for adding songs to the jukebox in
the future, by adding new genres of music to the selection choice.
All genres of music now have at the *least* five songs in them, so that
we no longer look like we've just left the jukebox as an afterthought.
(alt title: DJ Drof's Mixtape Volume 1, with Crew Input)