KMine
## About The Pull Request

Adding a Mining dogborg to the list of dogborgs since I found it odd that Cargo was the only department without one.

Functionally the same as a normal mining borg, save for the Supply Satchel belly module which contain the reagents:

   - Glucose
   - inaprovaline
   - tricordrazine

## Why It's Good For The Game

As mentioned before, Cargo was the only department without a dogborg model, so adding that in along with its own belly feature.

## Changelog
:cl:
N/A
/:cl: