Moves 'tie' accessories to vox species sheet
Because of the change in the override in accessories.dm, and with how most of the vox specific accessories such as webbing weren't defined an override, they needed to be moved over. Also moved a various amount of accessories that currently don't have vox fitting sprites, but will be fixed in a PR down the line.