F
uck is a profane English-language word[1] which often refers to the act of sexual intercourse but is also commonly used as an intensifier or to denote disdain. While its origin is obscure, it is usually considered to be first attested to around 1475. In modern usage, the term fuck and its derivatives (such as fucker and fucking) can be used as a noun, a verb, an adjective, an interjection or an adverb. There are many common phrases that employ the word as well as compounds that incorporate it, such as motherfucker, fuckwit, fuckup and fucknut.