Closes #2342
Humans use soap, not silicons.
Silicons are able to use mops, however.
Therefore, the soap that is given to drones (and only drones, no doubt an ancient unfixed strand of spaghetti) must be replaced by a mop to restore their ability to clean floors.

They must wet the mop in a sink, as any other synthetic or humanoid, then use it as normal.