removes unusable emote
i dont care enough to rewrite this to be more legible, so let's just be rid of it and call it a day. afaik, id have to add an else statement to every if because its a switch.... and im not doing that. removing it fixes it. works fine. o7