Changes to engine templates
## About The Pull Request
Adjusts the engine templates so that certain bugs are no longer present. Namely, everything wrong with the R-UST and duplicate radiator windows. Now it isn't just the SM that has radiator window shutters, as well.

## Why It's Good For The Game
Now, I know there was a ban on changing the layout of the map, so changing the layout of the pipes in the R-UST might be a bit of a stretch, but hear me out. Fixing the R-UST by means that you rightfully shouldn't need to take ICly every round fucking sucks.
The other things are just a bit of icing on the cake.
## Changelog
🆑Auris456852
fix: R-UST's hot loop now is set in the right direction and the wiring works.
fix: No more duplicate borosilicate engine windows.
tweak: Every engine submap now has radiator window shutters.
tweak: Removes the rogue button in the middle of the R-UST pipes.
/🆑