Part 1/4 of my update to explo/cargo
Splitting this up into four parts, because i'm super skittish about touching the shuttle points, and this is a good break spot anyways- This moves the QM and PF's offices, adds a 'secret' escape route, fills in the old space with some nice things, and as a bonus, fixes three other map issues on A1 while I was at it. ((A wall/door that lead to nothing, a timeclock terminal, and a stray pipe.))

Unlike the later PRs, which is why I'm doing the split here, this keeps the departments as closed up as they where before, and thus, I don't imagine there'll be any fuss over the changes. Maybe this'll also get me a maintainer actuially asking what the fuck I'm planning, so I can point them to the dozens of times I asked them specifically if what I'm planning was gonna fly or not. We'll see!