Part 1/4 of my update to explo/cargo (#3341)
* Part 1/4 of my update to explo/cargo

Splitting this up into four parts, because i'm super skittish about touching the shuttle points, and this is a good break spot anyways- This moves the QM and PF's offices, adds a 'secret' escape route, fills in the old space with some nice things, and as a bonus, fixes three other map issues on A1 while I was at it. ((A wall/door that lead to nothing, a timeclock terminal, and a stray pipe.))

Unlike the later PRs, which is why I'm doing the split here, this keeps the departments as closed up as they where before, and thus, I don't imagine there'll be any fuss over the changes. Maybe this'll also get me a maintainer actuially asking what the fuck I'm planning, so I can point them to the dozens of times I asked them specifically if what I'm planning was gonna fly or not. We'll see!

* Two quick fixed

Deck six is missing a single floor tile, deck five has a door over a wall instead of over a floor. Plus, took the chance to better name the doors.

* Fixes the access of boxes and lockers in lockup

Now people besides the captain can actuially use the gun cabinet in there, and the security weapons crates are locked to security only, so security can hold things in there.

* Getting fucky with cargo

Old cargo: moved almost entirely northwards, using the lower facing of the QM's window as a line. Everything interior adjusted accordingly. Pipes unscrambled somewhat. Planning placed out in Explo (three units of tape, a moved plant, and some windows) as well as markering off with construction tape the old cargo place. Vending hall is gone- vending machines scattered in construction space. When I move all the docking points in part 2, the construction space will be turned into an external trader's shop, all the machines organized so that visiting traders can use them,  and more.

* Oops. Fixes some airless plating where it doesn't belong

* Emergency patch, to try and fix shuttle and a pair of wires

* I FUCKED UP

* Hopfully fixes cargo shuttle

* Improves visuals, better hallway, better lobby

Not final form, but a big step in the right direction. More feedback on design needed