fixed 'sticky limbs' bug in character creation
altered copy_to_mob() in 03_body to check for whether it needs to generate a new organ. if it does (amputated limb to normal or robotic, changing robotic limb model, or robotic or amputated limb to normal) then it removes the limb and regenerates it
this required changing the order in which it goes through the limbs so that it won't try to, ie, put a hand back onto a missing arm, and this doesn't seem to have any apparent side-effects.

there's almost certainly a better solution, but this bug has existed for almost a year now and makes creating an FBP character an exercise in frustration.