Trap has been sprung.
Adds functions to the /obj/item/rig types for trapping.
If it's trapped, it will lock when you fully deploy and activate it. You will be unable to take it off or toggle parts.
If it's springtrapped, it will lock as above, but also deal 50 damage to your chest, and 70 spread through the rest of your body.

Also adds old, trapped subtypes for the current possessed mobs, which it drops either only trapped or a springtrapped module. No more risk free stuff.