The Possessed (#3484)
* They Come

Brings in the new faction type, The Possessed.

* The Possessed Breathe

Messed with the breathing code by accident. Working again. Oops.

* Noisy.

Adds footsteps, along with more death messages.

* Update the .dme

Woops. Apparently this is a thing I forgot to do.

* Trying to fix things.

* Delete possessed.dm

* Parent call.

* Miasmatic update.

Adds Miasma as a reagent, as well as a "Miasma bomb"  A chemical grenade that automatically detonates into a puff of smoke with miasma in it. And makes the possessed immune to grabs.

* Trap has been sprung.

Adds functions to the /obj/item/rig types for trapping.
If it's trapped, it will lock when you fully deploy and activate it. You will be unable to take it off or toggle parts.
If it's springtrapped, it will lock as above, but also deal 50 damage to your chest, and 70 spread through the rest of your body.

Also adds old, trapped subtypes for the current possessed mobs, which it drops either only trapped or a springtrapped module. No more risk free stuff.

* Fix the suit.

Fixes the issue where cutting off a trapped suit makes the suit stay on, even if the module is off. Also makes a trapped suit upgrade to springtrapped when cut off. Does nothing to normal rigs.

* Pain.

Made the springtrap a bit more painful.

* Better Springtrap.

The trap delays by 30 seconds default before it tries to activate. When 30 seconds pass, it tries to force deploy then activate, else it tries again in 30 seconds.

* Update rig.dm