Creates new item type, cleans up code, fixes unintentional buff (#3797)
* Creates new item type, cleans up code, fixes unintentional buff

Speak 'n' Rock headsets are headphones and Radio Headsets in one. Currently, they're not accessable via loadout. I have re-made my donator item as a subtype of this, to powercreep it a little later. That code is a lot more difficult then this is.

Speaking of difficult coding, Sillycons accidentally didn't adjust a thing, and all of the single-eared musical note headphones he craved to have on without sacrificing the radio earpiece where giving top-tier flashbang protection. this makes it so that they give none, as a stop-gap until further in-depth discussion can happen on the subject. ((Earmuffs are unaffected, and still give good flashbang protection))

* Fixes missing trailing endlines

because apparently an empty line at the end of things is important.

* Bad Tab

* Bad Spaces

Co-authored-by: Zandario <zandarioh@gmail.com>