Adds More Guns to Armory & Removes Erroneous Holster Flag for Warden's Shotgun (#4808)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

1. **Adds Guns to Armory.**
2. **Removes Holster Flag from Warden's Shotgun.**

## Why It's Good For The Game

1. _Adds eight spare tasers and four more LTL gunboxes to the Armory._
3. _This is apparently my fault. It appears to have come over with the
initial port, but it's possible I assigned the flag myself. Either way,
it's not an intended feature and I'm surprised it took this long for
anyone to mention it._

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Adds more guns to the Armory.
del: Removes Warden's Shotgun's holster flag.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->