Fix species display in character directory (#4938)
## About The Pull Request

Somehow during the reorg some of the necessary bits that make species
display properly in the chardir got eaten. This PR restores the
necessary .dm-side code to make it work again.

## Why It's Good For The Game

Fixes a silly bug so there's no longer a useless blank column in the
directory. Tested and works.

## Changelog

:cl:
fix: fixed species not displaying in character directory
/:cl: