Operation Waterlogged Sparrow (Introduction of Pirate Ghost Role) (#4909)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

1. **Fixes Fore Dir of Pirate Skiff.**
2. **Begins overhauling Vox Pirate Base Map.**

## Why It's Good For The Game

1. _The Pirate Skiff cannot be flown currently due to errors in its
code. This corrects that breakage._
2. _This Z-Level will be repurposed into the ghostrole base. I may
eventually delete this Z-level entirely and move the pirate base to
another Z-level to try and save on space, though that currently doesn't
make any sense. For now it's gonna stay as-is. At the time of writing,
it's been gutted of mobs and I'm just staging the redraw._

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Fixes Pirate Skiff fore dir.
tweak: Begins redraw of Vox Pirate Base into Pirate Base.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->