Adds a symbol replacement option for languages instead of syllables, adds Keisani language (#5074)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a Keisani specific language selectable by **anyone**. The general
idea is that the amount of of Keisanis has reached enough of a mass to
maybe kind of warrant this idk. The symbols aren't really necessary just
an exotic thing.

![image](https://user-images.githubusercontent.com/11361525/219846835-3e23a11e-c4a1-40d2-9ff1-e2a53f2acde2.png)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Also adds a way for any language to be ciphered with a list of symbols
## Why It's Good For The Game
Makes the general resurgence of Keisani a bit more meaningful to other
people who might care. Idk it's a language we have buckets of those.

Makes languages more fun
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:freestylalt
add:Ability to cipher languages using any symbols
add:Keisani language
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: silicons <2003111+silicons@users.noreply.github.com>