RASC Spectre Robotic/Prosthetic Body (Updated Version) (#5050)
## About The Pull Request

This PR includes the code and sprites for a new character selectable
Robotic/Prosthetic Body. that I ported from Splurt and then adjusted to
fit our codebase. Adjustments can be made to brand, name or description
if wished.

New PR version with updated lore naming and file directory to not cause
any issues and taking in previous feedback.

## Why It's Good For The Game

Adds extra flavour to the robotic/prosthetic choices for the players to
include a chassis ideally used for security and exploration duties.

## Changelog

:cl:
add: Adds Hoffman Tech RACS Spectre Robotic/Prosthetic Body 
imageadd: Adds a the sprites for the required RASC Spectre
/:cl:

---------

Co-authored-by: silicons <no@you.cat>