Initial QOL changes for mirrors and their respective tools. (#5114)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

it's on the tin and in the CL

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I like QOL.

In short, this updates the description of the tool and console to note
that you can use the tool directly on the console to transfer a mirror.
This was always the case, but now it's easier to tell. It also allows
you to load or unload mirrors the old way, IE using it in hand, and now
also allows you to click the tool with a mirror to load it like a weapon
magazine. I also made it so using the tool on a console won't slap the
console, which means the feedback for loading that has existed for a
year and a half that no one seems to notice is a bit more noticeable.
Also adds a note in the resleeving room with basic instructions.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
tweak: the mirror tool can now be loaded/unloaded like you would a gun
magazine.
qol: updated descriptions for the tool and console to make it clearer
that you can interface the two
fix: fixed the mirror tool and implants smacking things while being used
to make it a bit clearer what you're doing
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->