Adds a dice roll emote (#5008)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Adds the *roll emote. Using the params, you can roll a die of your
choice, add a bonus to your roll and compare it against a difficulty
check.

To use this emote specify what die you want to roll in the first line.
Prefix it with a "D". You may add a bonus to your roll, simply add a "+"
and your number after your chosen die. If you want to incorporate a
difficulty check, add a "-DC" at the end. You may use a bonus, a
difficulty check or both.

Here is an example: roll-D20+5-DC15. This will roll a twenty-sided die,
add five and compare it against a difficulty check of 15.


![image](https://user-images.githubusercontent.com/18533061/214977097-aec413c6-c451-4b0f-b489-847f4569b4f3.png)
(Typo already fixed, not getting a new image.

## Why It's Good For The Game

It's great for text RP to be able to fairly decide contests.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: *roll
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->