Disables most of modular weaponry. (#5173)
## Why it's good for the game 

Sorry.

Science fun, but science needs a rework.

It's at the point where I have to choose between science having fun with
their more than 200 DPS weapons, or the rest of the server not living
under dealing with this.

Science has more guns, on that note, that are actually just as bad or
worse.
I'm not touching them until they're clearly an issue, as promised. With
modular weapons, it's clearly an issue already.

This removal isn't permanent. I think a guncrafting system is cool.
I just don't feel leaving this in is a healthy thing to do anymore. It
isn't an administrative matter either, because if a powerful item
exists, science *exists* to hand it out.