Blueshields (#5155)
## About The Pull Request

Adds Blueshields, a personal bodyguard for the Director, and then other
heads/command when a Director is absent.


## Why It's Good For The Game

Soon we'll be on the shipmap, where exploration is crunched down to only
the pathfinder. With security becoming larger when the explorer mains
(likely) start switching, a personal guard for the high ranking is
requires so that the head of security can still manage the crew, and
remaining security officers while an away mission is taking place.
(Especially during events.)

The Blueshield is there to guard the Director first and foremost, and
they are *not* to play the role of a security officer unless instructed
to do so in the absence of a replacement. Security officers arriving
on-scene should take priority of apprehending suspects, while the
Blueshield's focus stays on moving the Director (or other head) to
safety.

Imo, staff should be lenient with the above guidelines when there are NO
other security members on board, or no command.

Finally, this is only my initial plan for how they are implemented.
Gameplay guidelines are free for editing when implemented into SOP, and
I'm happy to provide my input if pinged about it in the discord forum
post.

## Changelog

:cl:
add: Adds blueshields, a personal bodyguard for the Director, and other
heads/command when a Director is absent.
/:cl:

---------

Co-authored-by: silicons <no@you.cat>