Update Shadekin energy regenerations (#5196)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Shadekin energy modifiers have been set up inversed. The Values for
Light and for Dark were mixed.
In addition, some of the recharge values were pure CBT, which is not
needed for an adminspawn role.

## Why It's Good For The Game

This way the regeneration rates actually fit with the description of the
traits. Also, it removes unnecessary CBT, which is always good.

## Changelog

:cl:
fix: fixed Shadekin energy regen inversion
tweak: tweaked Shadekin energy regeneration
/:cl: