Greyscaled Human Torso Touchup (#5208)
The gore was sprited onto the torso, which caused issues with certain
limbs (vassilian) which had thinner portions at the root. This resulted
in pixels that appeared almost at random. I dont know why gore isnt an
overlay or something when a limb is removed, but this works *for now*.
Both male and female versions were touched up to remove dangling gore
pixels.

![before](https://user-images.githubusercontent.com/86886852/226201050-a4df2bba-48e1-4250-afbb-970a3bfdd989.PNG)

![after](https://user-images.githubusercontent.com/86886852/226201051-cff48d4d-c9be-42ca-8cb7-1b3524007c52.PNG)