Changes Gygax sprite to the 'advance' one (#5232)
Seems kinda silly this was just sitting in the code and not being used
anywhere, especially since it looks much better than the one currently
used.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Gives the Gygax a face lift by replacing the icon_state, initial_icon,
and wreckage lines with gygax_adv and mecha_wreckage/gygax/advance
respectively

## Why It's Good For The Game

Gives the Gygax a new fancy look instead of the currently used one. Just
found it sitting in the code unused anywhere else.

## Changelog


:cl:

imageadd: gygax_adv replaced instances of the original Gygax sprite,
same with the darkgygax.
imagedel: the original ones are simply swapped with the advance.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->