Holsters and Rigs (#5359)
Adds holsters to the loadout's accessory menu, and adds matrix
recolorable/renamable rigs to the utility section.

Yes the bug with machete holsters spawning filled was fixed, ping me if
people somehow find a way to spawn with them since it should be
impossible. Would likely be a result of some witchcraft, and very
intentional abuse of the loadout system.