Multitool Sprite Replacement (#5357)
## About The Pull Request

Replaces existing multitool sprites (red/green overlays too) with a new
sprite.

## Why It's Good For The Game

I want the server to lean away from the way default ss13 (tg and
similar) looks and feels as a whole. With the new wall/window/door
update zandario is doing, I'm looking to push sprite/sound upgrades
frequently to hopefully bring a fresh, unique feeling that doesn't
reflect base TG as much. Predicting that going this path of separating
our visual design from baseline TG will make new and returning players
enter with the assumption that our lore is unique, and start seeing
it/assuming it to be more than the current Nt vs syndicate (rebranded to
mercs) as the lore team (hopefully) continues to develop it further away
from that.

This will likely be the first of many bite-sized sprite/sound changes as
I further this goal in my free time.