Impromptu Living Statue Rework (#5416)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Just was supposed to be a port of tgstation/tgstation#73630, but I felt
like fixing the other things.

- Their spells function and are added.
- Their main gimmick is now a component, so that can be reused for other
objects/mobs.

## Changelog


:cl:
refactor: Living Statues properly function.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->