Made HE pipes 20000x as sensitive (#5415)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

As title. HE pipes now stop working if they're less than 0.01K off from
the environment instead of less than 20K.

## Why It's Good For The Game

HE pipes that only work with a less-than-20K delta are _completely
worthless_ for any actual, practical use. Imagine a heater that only
works if you're above 40C or below 0C; that is _literally_ what HE pipes
are right now. Even worse, it's not hysteresis, they _stop working_ if
they get within that range. It's really, really bad.

One might be concerned about performance issues, of course; well, I
tested this, live, on the server, only vaguely informing everyone that
it would happen, and measured the effects in profiling and the MC. It
had no overall effect on performance, and only a marginal effect in the
profiler even though I had, earlier in that round, set up a huge HE pipe
loop in maint connected directly to distro and even heated up distro a
bit _on purpose_ to make it work more. This is not a real performance
concern and the huge minimum-temp-delta is a major overcorrection.

## Changelog


:cl:
tweak: HE pipes now work with at least a 0.01K kelvin difference instead
of at least a 20K kelvin difference
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->